+++ inst/BNR/power/00.setup.R
@@ -12,8 +12,8 @@ pi0s <- c(0.8, 0.9, 0.99, 0.999)
 alphas <- c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.4, 0.5)
 
 m <- 1e3
-B <- 2e3
-nbSimu <- 2e3
+B <- 1e3
+nbSimu <- 1e3
 
 pname <- sprintf("m=%s,B=%s,nbSimu=%s", m, B, nbSimu)
 resPath <- "resData"