+++ inst/BNR/power/03.plotPowerVsJFWER.R
@@ -13,25 +13,27 @@ dat %<-% {
     plyr::ldply(fls, readRDS, .id="id")
 }
 head(dat)
+
 kc <- as.character(dat$kMax)
 kc[which(dat$kMax==m)] <- "m"
 kc[which(dat$kMax==m/2)] <- "m/2"
 kc[which(dat$kMax==(1-dat$pi0)*m)] <- "(1-pi0)*m"
 dat$kMaxC <- kc
-dat$estPowM1 <- dat$estPow/(1-dat$pi0)
+##dat$estPowM1 <- dat$estPow/(1-dat$pi0)
 
 powerz <-  list("Power1"="P(S(R,H1)>1", "Power"="P(S(R,H)>1", "s1"="E(S(R,H1))/m1", "s1"="E(S(R,H1))/m1")
 ##                res <- rbind(JR=rej0>0, detPow1=rej1>0, detPow=rej01>0, v0, estPow1=s1, estPow=s01)
 powerz <-  list("detPow1"="P(S(R,H1)>1", "detPow"="P(S(R,H)>1", 
-                "estPow1"="E(S(R,H1))/m1", "estPow"="E(S(R,H))/m", "estPowM1"="E(S(R,H))/m1")
+                "estPow1"="E(S(R,H1))/m1", "estPow"="E(S(R,H))/m1",
+                "powBH"="half of BH(0.05)", "pow0"="half of {p <= 0.05}")
 
 ## some reshaping
-datC <- subset(dat, kMax %in% as.character(c(10, 100, m, m/2))  & flavor != "unadjusted")
+#datC <- subset(dat, kMax %in% as.character(c(10, 100, m, m/2))  & flavor != "unadjusted")
 datC <- subset(dat, flavor != "unadjusted")
 
 mm <- grep("Oracle|alpha", datC$flavor)
 datC <- datC[-mm, ]
-levels(datC$family) <- list("Balanced"="kFWER", "Simes"="Simes")
+levels(datC$family) <- list("Balanced"="kFWER", "Linear"="Simes")
 datC$alpha <- as.numeric(datC$alpha)
 alphas <- unique(datC$alpha)
 datC$ff <- factor(paste(datC$family, datC$flavor))
@@ -51,7 +53,7 @@ pname2 <- gsub("0\\.", "", pname)  ## to avoid '.' in LaTeX file names
 ## plot Power vs JFWER
 library("ggplot2")
 
-x <- c("JR", "alpha")
+x <- c("JR", "alpha")[2]
 SNRs <- unique(dat$SNR)
 #SNRs <- 2